+++ modules/hac/ui/src/sap/commerce/toolset/hac/ui/HacConnectionSettingsDialog.kt
@@ -86,7 +86,7 @@ class HacConnectionSettingsDialog(
             sslProtocol = sslProtocolComboBox.selectedItem?.toString() ?: "",
             webroot = webrootTextField.text,
             timeout = timeoutIntSpinner.number,
-            sessionCookieName = sessionCookieNameTextField.text.takeIf { !it.isNullOrBlank() } ?: ExecConstants.DEFAULT_SESSION_COOKIE_NAME,
+            sessionCookieName = sessionCookieNameTextField.text.takeUnless { it.isNullOrBlank() } ?: ExecConstants.DEFAULT_SESSION_COOKIE_NAME,
         ),
         mutable.username.get(),
         mutable.password.get(),
@@ -144,34 +144,30 @@ class HacConnectionSettingsDialog(
             }
         }
 
-        group("Host Settings") {
+        collapsibleGroup("Host Settings") {
             row {
-                label("Address:")
                 hostTextField = textField()
-                    .comment("Host name or IP address")
+                    .label("Host / IP:")
                     .align(AlignX.FILL)
                     .bindText(mutable::host)
                     .onChanged { urlPreviewLabel.text = generateUrl() }
                     .addValidationRule("Address cannot be blank.") { it.text.isNullOrBlank() }
                     .component
-            }.layout(RowLayout.PARENT_GRID)
 
-            row {
-                label("Port:")
-                portTextField = textField()
-                    .align(AlignX.FILL)
+                portTextField = intTextField(1..65535)
+                    .label("Port:")
                     .bindText(mutable::port.toNonNullableProperty(""))
                     .onChanged { urlPreviewLabel.text = generateUrl() }
-                    .addValidationRule("Port should be blank or in a range of 1..65535.") {
-                        if (it.text.isNullOrBlank()) return@addValidationRule false
-
-                        val intValue = it.text.toIntOrNull() ?: return@addValidationRule true
-                        return@addValidationRule intValue !in 1..65535
-                    }
                     .component
             }.layout(RowLayout.PARENT_GRID)
 
             row {
+                webrootTextField = textField()
+                    .label("Webroot:")
+                    .bindText(mutable::webroot)
+                    .onChanged { urlPreviewLabel.text = generateUrl() }
+                    .component
+
                 sslProtocolCheckBox = checkBox("SSL:")
                     .bindSelected(mutable::ssl)
                     .onChanged { urlPreviewLabel.text = generateUrl() }
@@ -187,37 +183,26 @@ class HacConnectionSettingsDialog(
                 )
                     .enabledIf(sslProtocolCheckBox.selected)
                     .bindItem(mutable::sslProtocol.toNullableProperty())
-                    .align(AlignX.FILL)
                     .component
             }.layout(RowLayout.PARENT_GRID)
 
             row {
-                label("Webroot:")
-                webrootTextField = textField()
-                    .align(AlignX.FILL)
-                    .bindText(mutable::webroot)
-                    .onChanged { urlPreviewLabel.text = generateUrl() }
-                    .component
-            }.layout(RowLayout.PARENT_GRID)
-
-            row {
-                label("Session Cookie name:")
-                sessionCookieNameTextField = textField()
+                label("Session cookie:")
                     .comment("Optional: override the session cookie name. Default is JSESSIONID.")
-                    .align(AlignX.FILL)
+                sessionCookieNameTextField = textField()
                     .bindText(mutable::sessionCookieName)
                     .apply { component.text = "" }
                     .component
             }.layout(RowLayout.PARENT_GRID)
-        }
+        }.expanded = true
 
         if (SystemInfo.isWindows) {
-            group("Windows Subsystem for Linux") {
+            collapsibleGroup("Windows Subsystem for Linux") {
                 wslHostConfiguration()
-            }
+            }.expanded = true
         }
 
-        group("Authentication") {
+        collapsibleGroup("Authentication") {
             row {
                 segmentedButton(AuthenticationMode.entries.toList()) {
                     text = it.title
@@ -233,7 +218,14 @@ class HacConnectionSettingsDialog(
 
             authenticationAutomatic()
             authenticationManual()
-        }
+
+            separator()
+
+            row {
+                checkBox("Proxy authentication")
+                    .bindSelected(mutable.proxyAuthentication)
+            }
+        }.expanded = true
     }
 
     private fun Panel.authenticationManual() {